updated to work with new options interfaces
