+++ src/main/java/io/vertx/ext/sockjs/BridgeOptions.java
@@ -16,160 +16,58 @@
 package io.vertx.ext.sockjs;
 
 import io.vertx.codegen.annotations.Options;
-import io.vertx.core.json.JsonArray;
+import io.vertx.core.ServiceHelper;
 import io.vertx.core.json.JsonObject;
+import io.vertx.ext.sockjs.spi.BridgeOptionsFactory;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author <a href="http://tfox.org">Tim Fox</a>
  */
 @Options
-public class BridgeOptions {
+public interface BridgeOptions {
 
-  private static final long DEFAULT_AUTHTIMEOUT = 5 * 60 * 1000;
-  private static final String DEFAULT_AUTHADDRESS = "vertx.basicauthmanager.authorise";
-  private static final int DEFAULT_MAXADDRESSLENGTH = 200;
-  private static final int DEFAULT_MAXHANDLERSPERSOCKET = 1000;
-  private static final long DEFAULT_PINGTIMEOUT = 10 * 1000;
-  private static final long DEFAULT_REPLYTIMEOUT = 30 * 1000;
-
-  private long authTimeout;
-  private String authAddress;
-  private int maxAddressLength;
-  private int maxHandlersPerSocket;
-  private long pingTimeout;
-  private long replyTimeout;
-
-  private List<JsonObject> inboundPermitted = new ArrayList<>();
-  private List<JsonObject> outboundPermitted = new ArrayList<>();
-
-  public BridgeOptions() {
-    this.authTimeout = DEFAULT_AUTHTIMEOUT;
-    this.authAddress = DEFAULT_AUTHADDRESS;
-    this.maxAddressLength = DEFAULT_MAXADDRESSLENGTH;
-    this.maxHandlersPerSocket = DEFAULT_MAXHANDLERSPERSOCKET;
-    this.pingTimeout = DEFAULT_PINGTIMEOUT;
-    this.replyTimeout = DEFAULT_REPLYTIMEOUT;
+  static BridgeOptions options() {
+    return factory.options();
   }
 
-  public BridgeOptions(JsonObject json) {
-    this.authTimeout = json.getLong("authTimeout", DEFAULT_AUTHTIMEOUT);
-    this.authAddress = json.getString("authAddress", DEFAULT_AUTHADDRESS);
-    this.maxAddressLength = json.getInteger("maxAddressLength", DEFAULT_MAXADDRESSLENGTH);
-    this.maxHandlersPerSocket = json.getInteger("maxHandlersPerSocket", DEFAULT_MAXHANDLERSPERSOCKET);
-    this.pingTimeout = json.getLong("pingTimeout", DEFAULT_PINGTIMEOUT);
-    this.replyTimeout = json.getLong("replyTimeout", DEFAULT_REPLYTIMEOUT);
-    //TODO simplify common code
-    JsonArray arr = json.getArray("inboundPermitted");
-    if (arr != null) {
-      for (Object obj: arr) {
-        if (obj instanceof JsonObject) {
-          JsonObject jobj = (JsonObject)obj;
-          inboundPermitted.add(jobj);
-        } else {
-          throw new IllegalArgumentException("Invalid type " + obj.getClass() + " in inboundPermitted array");
-        }
-      }
-    }
-    arr = json.getArray("outboundPermitted");
-    if (arr != null) {
-      for (Object obj: arr) {
-        if (obj instanceof JsonObject) {
-          JsonObject jobj = (JsonObject)obj;
-          outboundPermitted.add(jobj);
-        } else {
-          throw new IllegalArgumentException("Invalid type " + obj.getClass() + " in outboundPermitted array");
-        }
-      }
-    }
+  static BridgeOptions optionsFromJson(JsonObject json) {
+    return factory.options(json);
   }
 
-  public long getAuthTimeout() {
-    return authTimeout;
-  }
+  long getAuthTimeout();
 
-  public BridgeOptions setAuthTimeout(long authTimeout) {
-    if (authTimeout < 1) {
-      throw new IllegalArgumentException("authTimeout must be > 0");
-    }
-    this.authTimeout = authTimeout;
-    return this;
-  }
+  BridgeOptions setAuthTimeout(long authTimeout);
 
-  public String getAuthAddress() {
-    return authAddress;
-  }
+  String getAuthAddress();
 
-  public BridgeOptions setAuthAddress(String authAddress) {
-    this.authAddress = authAddress;
-    return this;
-  }
+  BridgeOptions setAuthAddress(String authAddress);
 
-  public int getMaxAddressLength() {
-    return maxAddressLength;
-  }
+  int getMaxAddressLength();
 
-  public BridgeOptions setMaxAddressLength(int maxAddressLength) {
-    if (maxAddressLength < 1) {
-      throw new IllegalArgumentException("maxAddressLength must be > 0");
-    }
-    this.maxAddressLength = maxAddressLength;
-    return this;
-  }
+  BridgeOptions setMaxAddressLength(int maxAddressLength);
 
-  public int getMaxHandlersPerSocket() {
-    return maxHandlersPerSocket;
-  }
+  int getMaxHandlersPerSocket();
 
-  public BridgeOptions setMaxHandlersPerSocket(int maxHandlersPerSocket) {
-    if (maxHandlersPerSocket < 1) {
-      throw new IllegalArgumentException("maxHandlersPerSocket must be > 0");
-    }
-    this.maxHandlersPerSocket = maxHandlersPerSocket;
-    return this;
-  }
+  BridgeOptions setMaxHandlersPerSocket(int maxHandlersPerSocket);
 
-  public long getPingTimeout() {
-    return pingTimeout;
-  }
+  long getPingTimeout();
 
-  public BridgeOptions setPingTimeout(long pingTimeout) {
-    if (pingTimeout < 1) {
-      throw new IllegalArgumentException("pingTimeout must be > 0");
-    }
-    this.pingTimeout = pingTimeout;
-    return this;
-  }
+  BridgeOptions setPingTimeout(long pingTimeout);
 
-  public long getReplyTimeout() {
-    return replyTimeout;
-  }
+  long getReplyTimeout();
 
-  public BridgeOptions setReplyTimeout(long replyTimeout) {
-    if (replyTimeout < 1) {
-      throw new IllegalArgumentException("replyTimeout must be > 0");
-    }
-    this.replyTimeout = replyTimeout;
-    return this;
-  }
+  BridgeOptions setReplyTimeout(long replyTimeout);
 
-  public BridgeOptions addInboundPermitted(JsonObject json) {
-    inboundPermitted.add(json);
-    return this;
-  }
+  BridgeOptions addInboundPermitted(JsonObject json);
 
-  public List<JsonObject> getInboundPermitted() {
-    return inboundPermitted;
-  }
+  List<JsonObject> getInboundPermitted();
 
-  public BridgeOptions addOutboundPermitted(JsonObject json) {
-    outboundPermitted.add(json);
-    return this;
-  }
+  BridgeOptions addOutboundPermitted(JsonObject json);
 
-  public List<JsonObject> getOutboundPermitted() {
-    return outboundPermitted;
-  }
+  List<JsonObject> getOutboundPermitted();
+
+  static final BridgeOptionsFactory factory = ServiceHelper.loadFactory(BridgeOptionsFactory.class);
+
 }
