+++ src/main/java/io/vertx/ext/sockjs/impl/SockJSServerOptionsFactoryImpl.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2014 Red Hat, Inc.
+ *
+ *   Red Hat licenses this file to you under the Apache License, version 2.0
+ *   (the "License"); you may not use this file except in compliance with the
+ *   License.  You may obtain a copy of the License at:
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *   WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ *   License for the specific language governing permissions and limitations
+ *   under the License.
+ */
+
+package io.vertx.ext.sockjs.impl;
+
+import io.vertx.core.json.JsonObject;
+import io.vertx.ext.sockjs.SockJSServerOptions;
+import io.vertx.ext.sockjs.spi.SockJSServerOptionsFactory;
+
+/**
+ * @author <a href="http://tfox.org">Tim Fox</a>
+ */
+public class SockJSServerOptionsFactoryImpl implements SockJSServerOptionsFactory {
+
+  @Override
+  public SockJSServerOptions options() {
+    return new SockJSServerOptionsImpl();
+  }
+
+  @Override
+  public SockJSServerOptions options(JsonObject json) {
+    return new SockJSServerOptionsImpl(json);
+  }
+}
