+++ src/main/resources/META-INF/services/io.vertx.ext.sockjs.spi.BridgeOptionsFactory
@@ -1,0 +1,1 @@
+io.vertx.ext.sockjs.impl.BridgeOptionsFactoryImpl
