+++ src/main/resources/META-INF/services/io.vertx.ext.sockjs.spi.SockJSServerOptionsFactory
@@ -1,0 +1,1 @@
+io.vertx.ext.sockjs.impl.SockJSServerOptionsFactoryImpl
